Add files via upload
In this project python libraries like pillow, tkinter and pillow has been used. In this project the user can set a reminder for with a header and main message and cab schedule the time at which they need to be reminded or notified. They can clear their notifications according to their choice and can add them according to their preferences.